Relocated all synchronizer code from the controller to the synchronizer, and also moved synchonization onto its own thread.
