+++ src/main/java/org/qortal/controller/Controller.java
@@ -7,7 +7,6 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.security.SecureRandom;
 import java.security.Security;
 import java.time.LocalDateTime;
 import java.time.ZoneOffset;
@@ -52,7 +51,6 @@ import org.qortal.block.Block;
 import org.qortal.block.BlockChain;
 import org.qortal.block.BlockChain.BlockTimingByHeight;
 import org.qortal.controller.arbitrary.*;
-import org.qortal.controller.Synchronizer.SynchronizationResult;
 import org.qortal.controller.repository.PruneManager;
 import org.qortal.controller.repository.NamesDatabaseIntegrityCheck;
 import org.qortal.controller.tradebot.TradeBot;
@@ -94,14 +92,13 @@ public class Controller extends Thread {
 	public static final String VERSION_PREFIX = "qortal-";
 
 	private static final Logger LOGGER = LogManager.getLogger(Controller.class);
-	private static final long MISBEHAVIOUR_COOLOFF = 10 * 60 * 1000L; // ms
+	public static final long MISBEHAVIOUR_COOLOFF = 10 * 60 * 1000L; // ms
 	private static final int MAX_BLOCKCHAIN_TIP_AGE = 5; // blocks
 	private static final Object shutdownLock = new Object();
 	private static final String repositoryUrlTemplate = "jdbc:hsqldb:file:%s" + File.separator + "blockchain;create=true;hsqldb.full_log_replay=true";
 	private static final long NTP_PRE_SYNC_CHECK_PERIOD = 5 * 1000L; // ms
 	private static final long NTP_POST_SYNC_CHECK_PERIOD = 5 * 60 * 1000L; // ms
 	private static final long DELETE_EXPIRED_INTERVAL = 5 * 60 * 1000L; // ms
-	private static final long RECOVERY_MODE_TIMEOUT = 10 * 60 * 1000L; // ms
 	private static final int MAX_INCOMING_TRANSACTIONS = 5000;
 
 	// To do with online accounts list
@@ -114,7 +111,6 @@ public class Controller extends Thread {
 
 	private static volatile boolean isStopping = false;
 	private static BlockMinter blockMinter = null;
-	private static volatile boolean requestSync = false;
 	private static volatile boolean requestSysTrayUpdate = true;
 	private static Controller instance;
 
@@ -148,23 +144,8 @@ public class Controller extends Thread {
 	/** Whether we can mint new blocks, as reported by BlockMinter. */
 	private volatile boolean isMintingPossible = false;
 
-	/** Synchronization object for sync variables below */
-	private final Object syncLock = new Object();
-	/** Whether we are attempting to synchronize. */
-	private volatile boolean isSynchronizing = false;
-	/** Temporary estimate of synchronization progress for SysTray use. */
-	private volatile int syncPercent = 0;
-
 	/** List of incoming transaction that are in the import queue */
 	private List<TransactionData> incomingTransactions = Collections.synchronizedList(new ArrayList<>());
-
-	/** Latest block signatures from other peers that we know are on inferior chains. */
-	List<ByteArray> inferiorChainSignatures = new ArrayList<>();
-
-	/** Recovery mode, which is used to bring back a stalled network */
-	private boolean recoveryMode = false;
-	private boolean peersAvailable = true; // peersAvailable must default to true
-	private long timePeersLastAvailable = 0;
 
 	/** Lock for only allowing one blockchain-modifying codepath at a time. e.g. synchronization or newly minted block. */
 	private final ReentrantLock blockchainLock = new ReentrantLock();
@@ -354,20 +335,6 @@ public class Controller extends Thread {
 		return this.isMintingPossible;
 	}
 
-	public boolean isSynchronizing() {
-		return this.isSynchronizing;
-	}
-
-	public Integer getSyncPercent() {
-		synchronized (this.syncLock) {
-			return this.isSynchronizing ? this.syncPercent : null;
-		}
-	}
-
-	public boolean getRecoveryMode() {
-		return this.recoveryMode;
-	}
-
 	// Entry point
 
 	public static void main(String[] args) {
@@ -471,6 +438,9 @@ public class Controller extends Thread {
 				Controller.getInstance().shutdown();
 			}
 		});
+
+		LOGGER.info("Starting synchronizer");
+		Synchronizer.getInstance().start();
 
 		LOGGER.info("Starting block minter");
 		blockMinter = new BlockMinter();
@@ -582,11 +552,6 @@ public class Controller extends Thread {
 						// We can't do much without a valid NTP time
 						continue;
 					}
-				}
-
-				if (requestSync) {
-					requestSync = false;
-					potentiallySynchronize();
 				}
 
 				// Process incoming transactions queue
@@ -716,27 +681,6 @@ public class Controller extends Thread {
 		}
 	}
 
-	private long getRandomRepositoryMaintenanceInterval() {
-		final long minInterval = Settings.getInstance().getRepositoryMaintenanceMinInterval();
-		final long maxInterval = Settings.getInstance().getRepositoryMaintenanceMaxInterval();
-		if (maxInterval == 0) {
-			return 0;
-		}
-		return (new Random().nextLong() % (maxInterval - minInterval)) + minInterval;
-	}
-
-	/**
-	 * Export current trade bot states and minting accounts.
-	 */
-	public void exportRepositoryData() {
-		try (final Repository repository = RepositoryManager.getRepository()) {
-			repository.exportNodeLocalData();
-
-		} catch (DataException e) {
-			// Fail silently as this is an optional step
-		}
-	}
-
 	public static final Predicate<Peer> hasMisbehaved = peer -> {
 		final Long lastMisbehaved = peer.getPeerData().getLastMisbehaved();
 		return lastMisbehaved != null && lastMisbehaved > NTP.getTime() - MISBEHAVIOUR_COOLOFF;
@@ -761,7 +705,7 @@ public class Controller extends Thread {
 
 	public static final Predicate<Peer> hasInferiorChainTip = peer -> {
 		final PeerChainTipData peerChainTipData = peer.getChainTipData();
-		final List<ByteArray> inferiorChainTips = getInstance().inferiorChainSignatures;
+		final List<ByteArray> inferiorChainTips = Synchronizer.getInstance().inferiorChainSignatures;
 		return peerChainTipData == null || peerChainTipData.getLastBlockSignature() == null || inferiorChainTips.contains(new ByteArray(peerChainTipData.getLastBlockSignature()));
 	};
 
@@ -770,218 +714,34 @@ public class Controller extends Thread {
 		return peer.isAtLeastVersion(minPeerVersion) == false;
 	};
 
-	private void potentiallySynchronize() throws InterruptedException {
-		// Already synchronizing via another thread?
-		if (this.isSynchronizing)
-			return;
-
-		List<Peer> peers = Network.getInstance().getHandshakedPeers();
-
-		// Disregard peers that have "misbehaved" recently
-		peers.removeIf(hasMisbehaved);
-
-		// Disregard peers that only have genesis block
-		peers.removeIf(hasOnlyGenesisBlock);
-
-		// Disregard peers that don't have a recent block
-		peers.removeIf(hasNoRecentBlock);
-
-		// Disregard peers that are on an old version
-		peers.removeIf(hasOldVersion);
-
-		checkRecoveryModeForPeers(peers);
-		if (recoveryMode) {
-			peers = Network.getInstance().getHandshakedPeers();
-			peers.removeIf(hasOnlyGenesisBlock);
-			peers.removeIf(hasMisbehaved);
-			peers.removeIf(hasOldVersion);
+	private long getRandomRepositoryMaintenanceInterval() {
+		final long minInterval = Settings.getInstance().getRepositoryMaintenanceMinInterval();
+		final long maxInterval = Settings.getInstance().getRepositoryMaintenanceMaxInterval();
+		if (maxInterval == 0) {
+			return 0;
 		}
-
-		// Check we have enough peers to potentially synchronize
-		if (peers.size() < Settings.getInstance().getMinBlockchainPeers())
-			return;
-
-		// Disregard peers that have no block signature or the same block signature as us
-		peers.removeIf(hasNoOrSameBlock);
-
-		// Disregard peers that are on the same block as last sync attempt and we didn't like their chain
-		peers.removeIf(hasInferiorChainTip);
-
-		final int peersBeforeComparison = peers.size();
-
-		// Request recent block summaries from the remaining peers, and locate our common block with each
-		Synchronizer.getInstance().findCommonBlocksWithPeers(peers);
-
-		// Compare the peers against each other, and against our chain, which will return an updated list excluding those without common blocks
-		peers = Synchronizer.getInstance().comparePeers(peers);
-
-		// We may have added more inferior chain tips when comparing peers, so remove any peers that are currently on those chains
-		peers.removeIf(hasInferiorChainTip);
-
-		final int peersRemoved = peersBeforeComparison - peers.size();
-		if (peersRemoved > 0 && peers.size() > 0)
-			LOGGER.debug(String.format("Ignoring %d peers on inferior chains. Peers remaining: %d", peersRemoved, peers.size()));
-
-		if (peers.isEmpty())
-			return;
-
-		if (peers.size() > 1) {
-			StringBuilder finalPeersString = new StringBuilder();
-			for (Peer peer : peers)
-				finalPeersString = finalPeersString.length() > 0 ? finalPeersString.append(", ").append(peer) : finalPeersString.append(peer);
-			LOGGER.debug(String.format("Choosing random peer from: [%s]", finalPeersString.toString()));
-		}
-
-		// Pick random peer to sync with
-		int index = new SecureRandom().nextInt(peers.size());
-		Peer peer = peers.get(index);
-
-		actuallySynchronize(peer, false);
+		return (new Random().nextLong() % (maxInterval - minInterval)) + minInterval;
 	}
 
-	public SynchronizationResult actuallySynchronize(Peer peer, boolean force) throws InterruptedException {
-		boolean hasStatusChanged = false;
-		BlockData priorChainTip = this.getChainTip();
+	/**
+	 * Export current trade bot states and minting accounts.
+	 */
+	public void exportRepositoryData() {
+		try (final Repository repository = RepositoryManager.getRepository()) {
+			repository.exportNodeLocalData();
 
-		synchronized (this.syncLock) {
-			this.syncPercent = (priorChainTip.getHeight() * 100) / peer.getChainTipData().getLastHeight();
-
-			// Only update SysTray if we're potentially changing height
-			if (this.syncPercent < 100) {
-				this.isSynchronizing = true;
-				hasStatusChanged = true;
-			}
-		}
-		peer.setSyncInProgress(true);
-
-		if (hasStatusChanged)
-			updateSysTray();
-
-		try {
-			SynchronizationResult syncResult = Synchronizer.getInstance().synchronize(peer, force);
-			switch (syncResult) {
-				case GENESIS_ONLY:
-				case NO_COMMON_BLOCK:
-				case TOO_DIVERGENT:
-				case INVALID_DATA: {
-					// These are more serious results that warrant a cool-off
-					LOGGER.info(String.format("Failed to synchronize with peer %s (%s) - cooling off", peer, syncResult.name()));
-
-					// Don't use this peer again for a while
-					Network.getInstance().peerMisbehaved(peer);
-					break;
-				}
-
-				case INFERIOR_CHAIN: {
-					// Update our list of inferior chain tips
-					ByteArray inferiorChainSignature = new ByteArray(peer.getChainTipData().getLastBlockSignature());
-					if (!inferiorChainSignatures.contains(inferiorChainSignature))
-						inferiorChainSignatures.add(inferiorChainSignature);
-
-					// These are minor failure results so fine to try again
-					LOGGER.debug(() -> String.format("Refused to synchronize with peer %s (%s)", peer, syncResult.name()));
-
-					// Notify peer of our superior chain
-					if (!peer.sendMessage(Network.getInstance().buildHeightMessage(peer, priorChainTip)))
-						peer.disconnect("failed to notify peer of our superior chain");
-					break;
-				}
-
-				case NO_REPLY:
-				case NO_BLOCKCHAIN_LOCK:
-				case REPOSITORY_ISSUE:
-					// These are minor failure results so fine to try again
-					LOGGER.debug(() -> String.format("Failed to synchronize with peer %s (%s)", peer, syncResult.name()));
-					break;
-
-				case SHUTTING_DOWN:
-					// Just quietly exit
-					break;
-
-				case OK:
-					// fall-through...
-				case NOTHING_TO_DO: {
-					// Update our list of inferior chain tips
-					ByteArray inferiorChainSignature = new ByteArray(peer.getChainTipData().getLastBlockSignature());
-					if (!inferiorChainSignatures.contains(inferiorChainSignature))
-						inferiorChainSignatures.add(inferiorChainSignature);
-
-					LOGGER.debug(() -> String.format("Synchronized with peer %s (%s)", peer, syncResult.name()));
-					break;
-				}
-			}
-
-			// Has our chain tip changed?
-			BlockData newChainTip;
-
-			try (final Repository repository = RepositoryManager.getRepository()) {
-				newChainTip = repository.getBlockRepository().getLastBlock();
-			} catch (DataException e) {
-				LOGGER.warn(String.format("Repository issue when trying to fetch post-synchronization chain tip: %s", e.getMessage()));
-				return syncResult;
-			}
-
-			if (!Arrays.equals(newChainTip.getSignature(), priorChainTip.getSignature())) {
-				// Reset our cache of inferior chains
-				inferiorChainSignatures.clear();
-
-				Network network = Network.getInstance();
-				network.broadcast(broadcastPeer -> network.buildHeightMessage(broadcastPeer, newChainTip));
-			}
-
-			return syncResult;
-		} finally {
-			isSynchronizing = false;
-			peer.setSyncInProgress(false);
+		} catch (DataException e) {
+			// Fail silently as this is an optional step
 		}
 	}
 
-	private boolean checkRecoveryModeForPeers(List<Peer> qualifiedPeers) {
-		List<Peer> handshakedPeers = Network.getInstance().getHandshakedPeers();
-
-		if (handshakedPeers.size() > 0) {
-			// There is at least one handshaked peer
-			if (qualifiedPeers.isEmpty()) {
-				// There are no 'qualified' peers - i.e. peers that have a recent block we can sync to
-				boolean werePeersAvailable = peersAvailable;
-				peersAvailable = false;
-
-				// If peers only just became unavailable, update our record of the time they were last available
-				if (werePeersAvailable)
-					timePeersLastAvailable = NTP.getTime();
-
-				// If enough time has passed, enter recovery mode, which lifts some restrictions on who we can sync with and when we can mint
-				if (NTP.getTime() - timePeersLastAvailable > RECOVERY_MODE_TIMEOUT) {
-					if (recoveryMode == false) {
-						LOGGER.info(String.format("Peers have been unavailable for %d minutes. Entering recovery mode...", RECOVERY_MODE_TIMEOUT/60/1000));
-						recoveryMode = true;
-					}
-				}
-			} else {
-				// We now have at least one peer with a recent block, so we can exit recovery mode and sync normally
-				peersAvailable = true;
-				if (recoveryMode) {
-					LOGGER.info("Peers have become available again. Exiting recovery mode...");
-					recoveryMode = false;
-				}
-			}
-		}
-		return recoveryMode;
-	}
-
-	public void addInferiorChainSignature(byte[] inferiorSignature) {
-		// Update our list of inferior chain tips
-		ByteArray inferiorChainSignature = new ByteArray(inferiorSignature);
-		if (!inferiorChainSignatures.contains(inferiorChainSignature))
-			inferiorChainSignatures.add(inferiorChainSignature);
-	}
 
 	public static class StatusChangeEvent implements Event {
 		public StatusChangeEvent() {
 		}
 	}
 
-	private void updateSysTray() {
+	public void updateSysTray() {
 		if (NTP.getTime() == null) {
 			SysTray.getInstance().setToolTipText(Translator.INSTANCE.translate("SysTray", "SYNCHRONIZING_CLOCK"));
 			SysTray.getInstance().setTrayIcon(1);
@@ -997,13 +757,13 @@ public class Controller extends Thread {
 
 		String actionText;
 
-		synchronized (this.syncLock) {
+		synchronized (Synchronizer.getInstance().syncLock) {
 			if (this.isMintingPossible) {
 				actionText = Translator.INSTANCE.translate("SysTray", "MINTING_ENABLED");
 				SysTray.getInstance().setTrayIcon(2);
 			}
-			else if (this.isSynchronizing) {
-				actionText = String.format("%s - %d%%", Translator.INSTANCE.translate("SysTray", "SYNCHRONIZING_BLOCKCHAIN"), this.syncPercent);
+			else if (Synchronizer.getInstance().isSynchronizing()) {
+				actionText = String.format("%s - %d%%", Translator.INSTANCE.translate("SysTray", "SYNCHRONIZING_BLOCKCHAIN"), Synchronizer.getInstance().getSyncPercent());
 				SysTray.getInstance().setTrayIcon(3);
 			}
 			else if (numberOfPeers < Settings.getInstance().getMinBlockchainPeers()) {
@@ -1091,6 +851,9 @@ public class Controller extends Thread {
 						// We were interrupted while waiting for thread to join
 					}
 				}
+
+				LOGGER.info("Shutting down synchronizer");
+				Synchronizer.getInstance().shutdown();
 
 				// Export local data
 				LOGGER.info("Backing up local data");
@@ -1726,7 +1489,7 @@ public class Controller extends Thread {
 		peer.setChainTipData(newChainTipData);
 
 		// Potentially synchronize
-		requestSync = true;
+		Synchronizer.getInstance().requestSync();
 	}
 
 	private void onNetworkGetTransactionMessage(Peer peer, Message message) {
